Typography (#27)
* Update tokens

* Update colors palette story

* Fix minor grid layout issue

* Update typography, border-radius, spacing, box-shadow stories

* Add readme for foundations package

* Split color stories into separate modules

* Refactor themes story

* Remove unused css

* Use theme css variables when necessary

* Fix switch component

* Partially update playwright screenshots

* Update remaining failed screenshots

* Update theme token modificators mapping

* remove-refs

* Update foundations package README with ref token image

* Filter out tokens with "-ref-" in style-dictionary.mjs

* Update devDependencies in package.json

* Update devDependencies for parcel and TypeScript to latest versions

* Update Storybook dependencies to version 8.0.9

* Update Storybook dependencies to version 8.0.9

* Update CSS styles for toast and variableTooltip components

* Update CSS styles for toast and variableTooltip components

* remove ref token screenshots

* Refactor font weight handling in style-dictionary.mjs and dark.css

* Update Storybook color tests

* updated screenshots

* Removing garbage tokens from the pipeline

* Update font weight handling in style-dictionary.mjs and dark.css

* Update font family in base.css

* Merge duplicate selectors in merge-decl-plugin.js

* Update color-scheme in index.css and tokens-plugin.js

* Update Switch component in Storybook stories and switch.tsx

* Update TypeScript version in switch/package.json

* Update switch module CSS to change background color when checked

* Fix typo in foundations/README.MD

* Update colors and themes in Storybook CSS files

* Update typography package with new files and configurations

* chore: Update switch module CSS to use focus-visible pseudo-class

* feat: Add typography module CSS and update typography stories

The code changes include adding a new CSS module for typography and updating the typography stories in the Storybook. This will allow for better styling and customization of the typography components.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* feat: Add typography module CSS and update typography stories

* refactor: Update typography label font size handling

* feat: Add CSS module for typography and update typography stories

The code changes include adding a new CSS module for typography and updating the typography stories in the Storybook. This will allow for better styling and customization of the typography components.

* feat: Update typography components with CSS module and Storybook changes

* update screenshot

* chore: Update variableTooltip component to wrap children in a div

The code changes modify the variableTooltip component to wrap the children in a div element. This change ensures proper rendering and styling of the tooltip content.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* feat: Add label with switch to typography stories

The code changes include adding a new story to the typography component in the Storybook. This story demonstrates the usage of the label component with the switch component. It provides an example of how to combine these components for better user experience.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* chore: Update eslint rules for library.js

This commit updates the eslint rules for the library.js file. It adds new rule configurations for "import/no-default-export", "prefer-named-capture-group", and "unicorn/filename-case". These changes will help improve code quality and adhere to best practices.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* refactor: Update typography label font size handling

* refactor: Update typography label font size handling

* refactor: Update typography label font size handling

* refactor: Update Flex interface to use FlexSpacing for gap property

* refactor: Update switch module CSS to include cursor styles

The code changes in the switch module CSS add cursor styles to improve user interaction. The `:hover` state now changes the cursor to a pointer, and the `[data-disabled]` state changes the cursor to not-allowed. This enhances the usability of the switch component.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* refactor: Update switch module CSS to include cursor styles

* refactor: Update typography label font size handling

* refactor: Update typography label font size handling

* refactor: Update typography label color name

The code changes in the typography utils file update the color name from "tetriary" to "tertiary". This ensures consistency and clarity in the naming convention for text colors.

Co-authored-by: Igor Kornilov <gaber32188@gmail.com>

* refactor: Update typography label color name

The code changes in the typography utils file update the color name from "tetriary" to "tertiary". This ensures consistency and clarity in the naming convention for text colors.

* refactor: Update typography label color name

Update the typography label color name from "tetriary" to "tertiary" for consistency and clarity in the naming convention for text colors.

* chore: Update package manager to pnpm@9.1.0

---------

Co-authored-by: Tomas Hiltunen <tomas.hiltunen@gmail.com>